refactor: move components and hooks to special directories
