+++ src/components/offer-card-list/offer-card.tsx
@@ -1,8 +1,8 @@
 import {FC, MouseEventHandler} from 'react';
 import {Link} from 'react-router-dom';
-import {RoutePath} from '../../../shared/enums/routes.ts';
-import {FavoriteStatus} from '../../../shared/server-interaction/constants.ts';
-import {Offer} from '../model/types.ts';
+import {RoutePath} from '../../shared/enums/routes.ts';
+import {FavoriteStatus} from '../../shared/server-interaction/constants.ts';
+import {Offer} from '../../entities/offer/model/types.ts';
 
 type OfferCardProps = {
   offer: Offer;
