+++ src/pages/main-page/main-page.tsx
@@ -3,11 +3,11 @@ import {useSearchParams} from 'react-router-dom';
 import {CITY_SEARCH_PARAM} from '../../entities/city/model/constants.ts';
 import {City} from '../../entities/city/model/types.ts';
 import {isValidCity} from '../../entities/city/model/utils.ts';
-import {CityLinkList} from '../../entities/city/ui/city-link-list.tsx';
+import {CityLinkList} from '../../components/city-link-list/city-link-list.tsx';
 import {loadOffers, setCity} from '../../slices/offers-slice.ts';
-import {HeaderLogoLink} from '../../shared/components/header-logo-link.tsx';
+import {HeaderLogoLink} from '../../components/shared/header-logo-link.tsx';
 import {useAppDispatch, useAppSelector} from '../../shared/redux-helpers/typed-hooks.ts';
-import {HeaderUserInfo} from '../../widgets/common-components/header-user-info.tsx';
+import {HeaderUserInfo} from '../../components/shared/header-user-info.tsx';
 import {FullSpaceSpinner} from '../../widgets/spinner/ui/full-space-spinner.tsx';
 import {MainPageEmpty} from './main-page-empty.tsx';
 import {PlacesContainer} from './places-container.tsx';
