+++ src/pages/main-page/places-container.tsx
@@ -3,7 +3,7 @@ import {useNavigate} from 'react-router-dom';
 import {DEFAULT_CITY} from '../../entities/city/model/constants.ts';
 import {OfferSortOption} from '../../entities/offer/model/constants.ts';
 import {Offer} from '../../entities/offer/model/types.ts';
-import {OfferCardList} from '../../entities/offer/ui/offer-card-list.tsx';
+import {OfferCardList} from '../../components/offer-card-list/offer-card-list.tsx';
 import {changeFavoriteStatus} from '../../slices/favorites-page-slice.ts';
 import {setActiveOffer, setOffersSort} from '../../slices/offers-slice.ts';
 import {RoutePath} from '../../shared/enums/routes.ts';
