+++ src/pages/offer-page/offer-page-wrapper.tsx
@@ -1,6 +1,6 @@
 import {FC, ReactNode} from 'react';
-import {HeaderLogoLink} from '../../shared/components/header-logo-link.tsx';
-import {HeaderUserInfo} from '../../widgets/common-components/header-user-info.tsx';
+import {HeaderLogoLink} from '../../components/shared/header-logo-link.tsx';
+import {HeaderUserInfo} from '../../components/shared/header-user-info.tsx';
 
 export const OfferPageWrapper: FC<{children: ReactNode}> = ({children}) => {
   return (
